# cat: Improve performance on Linux