This PR should improve `cat` performance, especially on Linux. It does this by:

* Using trait monomorphization rather than trait objects (`Box<dyn Trait>`).
* Using `splice` on Linux

This PR supercedes https://github.com/uutils/coreutils/pull/1289

## Unscientific benchmarks

### GNU `cat`
```
[arni][~/src/coreutils][use-splice2]% cat /tmp/bigfile | pv -r > /dev/null
[4.47GiB/s]
[arni][~/src/coreutils][use-splice2]% cat /tmp/bigfile | pv -r > /dev/null
[4.68GiB/s]
[arni][~/src/coreutils][use-splice2]% cat /tmp/bigfile | pv -r > /dev/null
[4.65GiB/s]
```

### Before removing trait box, without Linux specialization
```
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[3.41GiB/s]
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[3.40GiB/s]
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[3.52GiB/s]
```

### After removing trait box, without Linux specialization
This is the figure you should expect for non Linux platforms.
```
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[7.76GiB/s]
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[5.67GiB/s]
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[6.61GiB/s]
```

### Linux specialization
```
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[29.6GiB/s]
--- 22:33:36 ---
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[14.4GiB/s]
--- 22:33:37 ---
[arni][~/src/coreutils][use-splice2]% target/release/cat /tmp/bigfile | pv -r > /dev/null
[22.2GiB/s]
```